Update CI jobs for libs and projects.
