+++ .github/workflows/airflow_jobs.yml
@@ -1,6 +1,6 @@
 # This is a basic workflow to help you get started with Actions
 
-name: CI
+name: projects/airflow
 
 # Controls when the workflow will run
 on:
@@ -22,8 +22,33 @@ on:
         default: false
 
 jobs:
-  build-and-unit-test:
+
+  code-quality:
     runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: ./projects/airflow
+
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.8
+      - name: Setup Poetry
+        uses: abatilo/actions-poetry@v2.0.0
+        with:
+          poetry-version: 1.1.11
+      - name: Install dependencies
+        run: poetry install
+      - name: Lint with flake8
+        run: poetry run flake8
+
+  unit-tests:
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: ./projects/airflow
 
     steps:
       - uses: actions/checkout@v2
@@ -39,16 +64,15 @@ jobs:
       - name: Build docker stack
         run: make setup
       - name: Wait for container health
-        run: bash wait-for-healthy-container.sh airflow-scheduler 180
-      - name: Lint with flake8
-        run: make lint
-      - name: Check code styles with black
-        run: make stylecheck
+        run: bash wait-for-healthy-container.sh airflow-scheduler 300
       - name: Run unit tests
         run: make unittest
 
   integration-tests:
     runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: ./projects/airflow
 
     steps:
       - uses: actions/checkout@v2
@@ -64,7 +88,7 @@ jobs:
       - name: Build docker stack
         run: make setup
       - name: Wait for container health
-        run: bash wait-for-healthy-container.sh airflow-scheduler 180
+        run: bash wait-for-healthy-container.sh airflow-scheduler 300
       - name: Run DB migrations
         run: make migrations
       - name: Setup tmate session
