+++ libs/database/Makefile
@@ -9,13 +9,13 @@ build:
 	 DOCKER_BUILDKIT=1 docker-compose build
 
 setup:
-	docker-compose up -d && make migrations
+	docker-compose up -d
 
 down:
 	docker-compose down -v
 
 unittest:
-	poetry run pytest -v -k 'not integration' -n 'auto'
+	poetry run pytest -v -n 'auto'
 
 lint:
 	poetry run flake8
@@ -25,6 +25,3 @@ stylecheck:
 
 watch:
 	poetry run pytest -f
-
-migrations:
-	poetry run yoyo apply --no-config-file --database postgresql://${TARGET_DB_USER}:${TARGET_DB_PW}@${TARGET_DB_HOSTNAME}/${TARGET_DB} ./dags/database/migrations/migration_files
