+++ libs/database/database/__init__.py
@@ -2,8 +2,11 @@ from .db_context import (
     create_db_context,
     teardown_db_context,
 )
-from .fixtures import db_context
-from .db_test_context import teardown_test_db_context, create_test_db_context
+from .db_test_context import (
+    teardown_test_db_context,
+    create_test_db_context,
+    db_context,
+)
 from .types import DBCredentials, DBContext
 from .execute_sql import (
     execute_values,
