+++ libs/database/database/conftest.py
@@ -1,3 +1,3 @@
-from .fixtures import db_context
+from .db_test_context import db_context
 
 __all__ = ["db_context"]
