+++ libs/database/pyproject.toml
@@ -2,7 +2,7 @@
 name = "database"
 version = "0.1.0"
 description = ""
-authors = ["Jakob J. Kolb <kolb@hardfork.io>"]
+authors = ["Jakob J. Kolb <jakob.j.kolb@gmail.com>"]
 
 [tool.poetry.dependencies]
 python = "^3.8"
