+++ projects/airflow/Makefile
@@ -21,16 +21,10 @@ integrationtest:
 	docker exec airflow-scheduler pytest -v -k 'integration'
 
 lint:
-	docker exec airflow-scheduler poetry run flake8 
-	
-stylecheck:
-	docker exec airflow-scheduler poetry run black --check
+	poetry run flake8
 
 watch:
 	docker exec -it airflow-scheduler pytest -f --ignore ./logs -k 'not integration'
-
-watchintegration:
-	docker exec -it airflow-scheduler pytest -k 'integration' -n 'auto' -f
 
 migrations:
 	docker exec -w /opt/airflow/ airflow-scheduler poetry run yoyo apply --no-config-file --database postgresql://${TARGET_DB_USER}:${TARGET_DB_PW}@${TARGET_DB_HOSTNAME}/${TARGET_DB} ./dags/database/migrations/migration_files
